Extract common scraper logic into abstract class
